+++ addon/mixins/regenerated/serializers/new-platform-flexberry-g-i-s-map-object-setting.js
@@ -1,11 +1,27 @@
+/**
+  @module ember-flexberry-gis
+*/
+
 import Ember from 'ember';
 
+/**
+  Mixin containing initializetion logic for map object setting serializer.
+
+  @class NewPlatformFlexberyGISMapObjectSettingSerializerMixin
+  @extends <a href="http://emberjs.com/api/classes/Ember.Mixin.html">Ember.Mixin</a>
+*/
 export let Serializer = Ember.Mixin.create({
-  attrs: {
-    defaultMap: { serialize: 'odata-id', deserialize: 'records' }
+  getAttrs: function () {
+    let parentAttrs = this._super();
+    let attrs = {
+      defaultMap: { serialize: 'odata-id', deserialize: 'records' }
+    };
+
+    return Ember.$.extend(true, {}, parentAttrs, attrs);
   },
-  /**
-  * Field name where object identifier is kept.
-  */
-  primaryKey: '__PrimaryKey'
+
+  init: function () {
+    this.set('attrs', this.getAttrs());
+    this._super(...arguments);
+  }
 });