+++ addon/models/new-platform-flexberry-g-i-s-layer-metadata.js
@@ -3,29 +3,30 @@
 */
 
 import Ember from 'ember';
-import {
-  Model as NewPlatformFlexberyGISLayerMetadataModelMixin,
-  defineProjections
-} from '../mixins/regenerated/models/new-platform-flexberry-g-i-s-layer-metadata';
+import { Model as LayerMetadataMixin, defineProjections } from '../mixins/regenerated/models/new-platform-flexberry-g-i-s-layer-metadata';
 import { Projection } from 'ember-flexberry-data';
+import { Offline } from 'ember-flexberry-data';
+import LayerModelMixin from '../mixins/layer-model';
+import LeafletCrsMixin from '../mixins/leaflet-crs';
 
 /**
-  Map layer metadata model.
+  Layer metadata model.
 
-  @class NewPlatformFlexberryGISLayerMetadata
+  @class NewPlatformFlexberryGISLayerMetadataModel
   @extends Model
-  @uses NewPlatformFlexberyGISLayerMetadataModelMixin
+  @uses OfflineModelMixin
+  @uses NewPlatformFlexberryGISLayerMetadataModelMixin
+  @uses LayerModelMixin
   @uses LeafletCrsMixin
 */
-let Model = Projection.Model.extend(NewPlatformFlexberyGISLayerMetadataModelMixin, {
-  settingsAsObject: Ember.computed('settings', function () {
-    let stringToDeserialize = this.get('settings');
-    if (stringToDeserialize) {
-      return JSON.parse(stringToDeserialize);
-    }
+let Model = Projection.Model.extend(Offline.ModelMixin, LayerMetadataMixin, LayerModelMixin, LeafletCrsMixin, {
+  _anyTextChanged: Ember.on('init', Ember.observer('name', 'description', 'keyWords', function() {
+    Ember.run.once(this, '_anyTextCompute');
+  })),
 
-    return {};
-  })
+  anyTextCompute() {
+    return `${this.get('name') || ''} ${this.get('description') || ''} ${(this.get('keyWords') || '').replace(',', ' ')}`;
+  }
 });
 
 defineProjections(Model);