+++ addon/serializers/new-platform-flexberry-g-i-s-link-parameter.js
@@ -1,6 +1,20 @@
+/**
+  @module ember-flexberry-gis
+*/
+
 import FlexberryData from 'ember-flexberry-data';
 import { Serializer as LinkParameterSerializer } from '../mixins/regenerated/serializers/new-platform-flexberry-g-i-s-link-parameter';
 
-export default FlexberryData.Serializer.Odata.extend(LinkParameterSerializer, {
+/**
+  Link parameter serializer.
 
+  @class NewPlatformFlexberryGISLinkParameterSerializer
+  @extends OdataSerializer
+  @uses NewPlatformFlexberryGISLinkParameterSerializerMixin
+*/
+export default FlexberryData.Serializer.Odata.extend(LinkParameterSerializer, {
+  /**
+  * Field name where object identifier is kept.
+  */
+  primaryKey: '__PrimaryKey'
 });