+++ app/controllers/new-platform-flexberry-g-i-s-layer-metadata-e/new.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-gis/controllers/new-platform-flexberry-g-i-s-layer-metadata-e/new';