+++ app/models/new-platform-flexberry-g-i-s-link-metadata.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-gis/models/new-platform-flexberry-g-i-s-link-metadata';